Fix docs and remove logic from Datable
